refactor(cli): Rename serve ACP HTTP directory
Rename the serve ACP HTTP module directory from acpHttp to acp-http and update the matching imports and current documentation paths.
