+++ docs/developers/daemon/02-serve-runtime.md
@@ -48,7 +48,7 @@
 | `serve/auth/device-flow.ts`, `qwen-device-flow-provider.ts`      | Device-flow OAuth routes. See [`12-auth-security.md`](./12-auth-security.md).                                                                                                                                                                                                                                                                                                                                                                                |
 | `serve/daemon-logger.ts`                                         | `DaemonLogger` structured file logs. See [`19-observability.md`](./19-observability.md).                                                                                                                                                                                                                                                                                                                                                                     |
 | `serve/debug-mode.ts`                                            | Shared `isServeDebugMode()` predicate controlling verbose error context in HTTP responses.                                                                                                                                                                                                                                                                                                                                                                   |
-| `serve/acpHttp/`                                                 | ACP Streamable HTTP transport (RFD #721), mounted at `/acp`. Seven files implement JSON-RPC POST, SSE GET, DELETE teardown, and shared bridge usage in parallel with the REST surface.                                                                                                                                                                                                                                                                       |
+| `serve/acp-http/`                                                | ACP Streamable HTTP transport (RFD #721), mounted at `/acp`. Seven files implement JSON-RPC POST, SSE GET, DELETE teardown, and shared bridge usage in parallel with the REST surface.                                                                                                                                                                                                                                                                       |
 | `serve/demo.ts`                                                  | Self-contained inline HTML for `GET /demo`: browser debug console with chat UI, event log, and workspace inspector. On loopback without `--require-auth`, it is registered **before** `bearerAuth`; on non-loopback or with `--require-auth`, it is registered **after** `bearerAuth`. Served with CSP `default-src 'none'; script-src 'unsafe-inline'; style-src 'unsafe-inline'; connect-src 'self'; frame-ancestors 'none'` plus `X-Frame-Options: DENY`. |
 
 **Re-export shims** for compatibility with pre-F1 import paths:
