+++ packages/cli/src/serve/acp-http/connection-registry.test.ts
