+++ packages/cli/src/serve/acp-http/sse-stream.test.ts
