+++ packages/cli/src/serve/acp-http/sse-stream.ts
